# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BriefingGenerateRequest": {
        "additionalProperties": false,
        "properties": {
          "briefing_date": {
            "format": "date",
            "title": "Briefing Date",
            "type": "string"
          },
          "local_time": {
            "pattern": "^\\d{2}:\\d{2}$",
            "title": "Local Time",
            "type": "string"
          },
          "regenerate": {
            "default": false,
            "title": "Regenerate",
            "type": "boolean"
          },
          "timezone": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "briefing_date",
          "local_time",
          "timezone"
        ],
        "title": "BriefingGenerateRequest",
        "type": "object"
      },
      "BriefingItemPublic": {
        "properties": {
          "article_id": {
            "title": "Article Id",
            "type": "integer"
          },
          "display_rank": {
            "title": "Display Rank",
            "type": "number"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "provenance_json": {
            "additionalProperties": true,
            "title": "Provenance Json",
            "type": "object"
          }
        },
        "required": [
          "article_id",
          "position",
          "display_rank",
          "provenance_json"
        ],
        "title": "BriefingItemPublic",
        "type": "object"
      },
      "BriefingPublic": {
        "properties": {
          "briefing_date": {
            "title": "Briefing Date",
            "type": "string"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "content_markdown": {
            "title": "Content Markdown",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "generation_version": {
            "title": "Generation Version",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BriefingItemPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "local_time": {
            "title": "Local Time",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "security_context": {
            "title": "Security Context",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "briefing_date",
          "local_time",
          "timezone",
          "title",
          "content_markdown",
          "status",
          "security_context",
          "generation_version",
          "revision",
          "checksum",
          "error_message",
          "items"
        ],
        "title": "BriefingPublic",
        "type": "object"
      },
      "BriefingSchedulePublic": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "local_time": {
            "title": "Local Time",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "local_time",
          "timezone",
          "is_active"
        ],
        "title": "BriefingSchedulePublic",
        "type": "object"
      },
      "BriefingScheduleUpdate": {
        "additionalProperties": false,
        "properties": {
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "local_time": {
            "pattern": "^\\d{2}:\\d{2}$",
            "title": "Local Time",
            "type": "string"
          },
          "timezone": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "local_time",
          "timezone"
        ],
        "title": "BriefingScheduleUpdate",
        "type": "object"
      },
      "ClusterItemPublic": {
        "properties": {
          "article_id": {
            "title": "Article Id",
            "type": "integer"
          },
          "membership_reason": {
            "title": "Membership Reason",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "similarity_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity Score",
            "type": "number"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "article_id",
          "title",
          "url",
          "source_id",
          "source_name",
          "summary",
          "tags",
          "published_at",
          "status",
          "similarity_score",
          "membership_reason"
        ],
        "title": "ClusterItemPublic",
        "type": "object"
      },
      "ClusterPublic": {
        "properties": {
          "algorithm_version": {
            "title": "Algorithm Version",
            "type": "string"
          },
          "curation_state": {
            "title": "Curation State",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClusterItemPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "reconciled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconciled At"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "window_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "window_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "status",
          "curation_state",
          "algorithm_version",
          "window_start",
          "window_end",
          "reconciled_at",
          "item_count",
          "source_count",
          "items"
        ],
        "title": "ClusterPublic",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Tuxnews",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/briefings": {
      "get": {
        "operationId": "list_briefings_api_v1_briefings_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BriefingPublic"
                  },
                  "title": "Response List Briefings Api V1 Briefings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Briefings",
        "tags": [
          "briefings"
        ]
      }
    },
    "/api/v1/briefings/generate": {
      "post": {
        "operationId": "generate_briefing_route_api_v1_briefings_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefingGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Briefing Route",
        "tags": [
          "briefings"
        ]
      }
    },
    "/api/v1/briefings/schedule": {
      "get": {
        "operationId": "get_briefing_schedule_api_v1_briefings_schedule_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingSchedulePublic"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Briefing Schedule",
        "tags": [
          "briefings"
        ]
      },
      "put": {
        "operationId": "put_briefing_schedule_api_v1_briefings_schedule_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefingScheduleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingSchedulePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Briefing Schedule",
        "tags": [
          "briefings"
        ]
      }
    },
    "/api/v1/briefings/today": {
      "get": {
        "operationId": "get_today_briefing_api_v1_briefings_today_get",
        "parameters": [
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "maxLength": 64,
              "minLength": 1,
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Today Briefing",
        "tags": [
          "briefings"
        ]
      }
    },
    "/api/v1/briefings/{briefing_id}": {
      "get": {
        "operationId": "get_briefing_api_v1_briefings__briefing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "briefing_id",
            "required": true,
            "schema": {
              "title": "Briefing Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Briefing",
        "tags": [
          "briefings"
        ]
      }
    },
    "/api/v1/briefings/{briefing_id}/regenerate": {
      "post": {
        "operationId": "regenerate_briefing_api_v1_briefings__briefing_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "briefing_id",
            "required": true,
            "schema": {
              "title": "Briefing Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Regenerate Briefing",
        "tags": [
          "briefings"
        ]
      }
    },
    "/api/v1/clusters": {
      "get": {
        "operationId": "list_clusters_api_v1_clusters_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClusterPublic"
                  },
                  "title": "Response List Clusters Api V1 Clusters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Clusters",
        "tags": [
          "clusters"
        ]
      }
    },
    "/api/v1/clusters/{cluster_id}": {
      "get": {
        "operationId": "get_cluster_api_v1_clusters__cluster_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "title": "Cluster Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Cluster",
        "tags": [
          "clusters"
        ]
      }
    },
...
  }
}